refactor util.py -> utils.py
