+++ incydr/_cases/client.py
@@ -15,7 +15,7 @@ from incydr._cases.models import CasesPage
 from incydr._cases.models import CreateCaseRequest
 from incydr._cases.models import QueryCasesRequest
 from incydr._cases.models import UpdateCaseRequest
-from incydr._core.util import get_filename_from_content_disposition
+from incydr._core.utils import get_filename_from_content_disposition
 from incydr._file_events.models.event import FileEventV2
 from incydr.enums import SortDirection
 from incydr.enums.cases import CaseStatus
